refactor: narrow the wire `channel` from `u32` to `u8`
Every codec decodes the frame header's `channel_id` as a `u8` and the
registry already stores `Option<u8>`, but `processor.rs` widened it with
`as u32` on the way into a `FeedMessage` for no stated reason, and that
width then propagated through the catalog key, `BookKey`/`MarketKey`, the
history key, the product id and the WebSocket subscribe filter. The
edge-feed-spec glossary pins Channel ID at `u8`; this makes the whole
crate agree with the wire and removes the widening casts along with it.

Emitted JSON is byte-identical: JSON numbers carry no width, so a
`channel` still renders as a bare number.

The one visible effect is on input. A `subscribe` frame carrying
`{"channel": 300}` no longer deserializes, so the client is answered with
the sink's `{"channel":"error","error":"unrecognized message"}` frame
instead of having the filter accepted and then matching nothing for the
life of the connection. The connection itself survives.

The arbiter's channel-clock eviction test drove `MAX_CHANNEL_CLOCKS + 8`
distinct channels, which only 256 values can no longer reach. It now
varies the venue instead: the clock key is `(venue, category, channel)`
and the venue is the wire Source ID's name, bounded only by
`MAX_UNREGISTERED_SOURCES`, so the cap still binds.

Refs #142

+++ doublezero-edge/src/channels.rs
@@ -89,7 +89,7 @@ pub struct DropRow {
     pub venue: String,
     pub category: String,
     pub code: String,
-    pub channel: u32,
+    pub channel: u8,
     pub name: String,
     pub products: u64,
 }
